backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0
pymysql==1.1.0
pyjwt==2.8.0
PyJWT==2.6.0